requirements.txt
fastapi
uvicorn
python-jose[cryptography]   # for JWT

